Moved Controllers to watchController.js
